RDA/requirements.txt

djangorestframework==3.14.0
sqlparse==0.4.3
asgiref==3.6.0
beautifulsoup4==4.12.0
Django==3.2.18
django-bootstrap-v5==1.0.11
django-gcharts==1.3
django-google-charts==0.1.1
django-qsstats-magic==1.1.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
soupsieve==2.4


